+++ packages/next/src/client.ts
@@ -1,4 +1,0 @@
-"use client";
-
-export * from "./sdk/client";
-export { CreateSdkContextReturn } from "./sdk/types";