+++ packages/sdk/src/modules/middlewareModule/utils/getRequestSender.ts
@@ -6,8 +6,10 @@ import {
   HTTPClient,
   ErrorHandler,
   RequestSender,
+  Logger,
 } from "../types";
 import { SdkHttpError } from "./SdkHttpError";
+import { defaultLogger } from "./defaultLogger";
 
 /**
  * Generates a `RequestSender` function configured according to the provided options.
@@ -22,7 +24,7 @@ export const getRequestSender = (options: Options): RequestSender => {
     ssrApiUrl,
     defaultRequestConfig = {},
     methodsRequestConfig = {},
-    cdnCacheBustingId,
+    cdnCacheBustingId = undefined,
   } = options;
 
   const getUrl = (
@@ -48,9 +50,11 @@ export const getRequestSender = (options: Options): RequestSender => {
     // If there are query params, append them to the URL as `?body=[<strignified query params>]`
     const serializedBody = encodeURIComponent(JSON.stringify(params));
     // Serialize CDN cache busting ID
-    const serializedCdnCacheBustingId = encodeURIComponent(cdnCacheBustingId);
+    const serializedCdnCacheBustingId = cdnCacheBustingId
+      ? `&cdnCacheBustingId=${encodeURIComponent(cdnCacheBustingId)}`
+      : "";
 
-    return `${url}?body=${serializedBody}&cdnCacheBustingId=${serializedCdnCacheBustingId}`;
+    return `${url}?body=${serializedBody}${serializedCdnCacheBustingId}`;
   };
 
   const getConfig = (
@@ -64,15 +68,25 @@ export const getRequestSender = (options: Options): RequestSender => {
       Accept: "application/json",
       ...defaultRequestConfig.headers,
     };
-    const mergedHeaders = {
-      ...defaultHeaders,
-      ...methodHeaders,
-      ...headers,
-    };
+    const mergedHeaders = Object.fromEntries(
+      Object.entries({
+        ...defaultHeaders,
+        ...methodHeaders,
+        ...headers,
+      }).reduce<[string, string | string[]][]>(
+        (headersSoFar, [key, value]) =>
+          // Removing Content-Lenght header if it equals 0 because it would crash Next.js server
+          // eslint-disable-next-line eqeqeq
+          key.toLocaleLowerCase() === "content-length" && value == "0"
+            ? headersSoFar
+            : [...headersSoFar, [key, value]],
+        []
+      )
+    );
 
     const computedHeaders: ComputedConfig["headers"] = {};
     Object.entries(mergedHeaders).forEach(([key, value]) => {
-      computedHeaders[key] = Array.isArray(value) ? value.join(",") : value;
+      computedHeaders[key] = Array.isArray(value) ? value.join(";") : value;
     });
 
     return {
@@ -82,6 +96,15 @@ export const getRequestSender = (options: Options): RequestSender => {
         ...computedHeaders,
       },
     };
+  };
+
+  const getLogger = (logger: Options["logger"]): Logger => {
+    if (logger === true) return defaultLogger;
+    if (logger === false) return {};
+    if (!logger)
+      return process.env.ALOKAI_SDK_DEBUG === "true" ? defaultLogger : {};
+
+    return logger;
   };
 
   const defaultHTTPClient: HTTPClient = async (url, params, config) => {
@@ -94,9 +117,14 @@ export const getRequestSender = (options: Options): RequestSender => {
     const responseJson = await response.json().catch(() => undefined);
 
     if (!response.ok) {
+      const message = `${config?.method} ${url} ${
+        responseJson?.message ?? response.statusText
+      }`;
+
       throw new SdkHttpError({
         statusCode: response.status,
-        message: responseJson?.message,
+        message,
+        url,
       });
     }
 
@@ -112,6 +140,7 @@ export const getRequestSender = (options: Options): RequestSender => {
       httpClient = defaultHTTPClient,
       errorHandler = defaultErrorHandler,
     } = options;
+    const logger = getLogger(options.logger);
     const { method, headers = {}, ...restConfig } = config ?? {};
     const methodConfig = methodsRequestConfig[methodName] || {};
     const finalMethod =
@@ -122,9 +151,24 @@ export const getRequestSender = (options: Options): RequestSender => {
       { method: finalMethod, headers, ...restConfig },
       methodConfig
     );
+    const startTime = performance.now();
+    logger.onRequest?.({
+      config: finalConfig,
+      params,
+      url: finalUrl,
+    });
 
     try {
-      return await httpClient(finalUrl, finalParams, finalConfig);
+      const response = await httpClient(finalUrl, finalParams, finalConfig);
+      const responseTime = performance.now() - startTime;
+      logger.onResponse?.({
+        config: finalConfig,
+        params,
+        response,
+        responseTime,
+        url: finalUrl,
+      });
+      return response;
     } catch (error) {
       return await errorHandler({
         error,