+++ packages/sdk/CHANGELOG.md
@@ -1,5 +1,57 @@
 # Change log
 
+## 3.2.3
+
+### Patch Changes
+
+- **[FIXED]** Headers with same name will now have their values merged using the ";' delimiter, rather than ",". This is seen when middleware and defaultRequestConfig set the same header - e.g. `Cookie`.
+
+After this change, `Cookie: vsf-locale=en` + `Cookie: custom-cookie=123` will result in `Cookie: vsf-locale=en; custom-cookie=123`, rather than being wrongly separated by `,`.
+
+## 3.2.2
+
+### Patch Changes
+
+- **[CHANGED]** response type of HTTPClient to `Promise<any>`. Previously it was `Promise<true> | SdkHttpError`, which was making it impossible the use of an `axios` with the `middlewareModule`. Now, it more flexible and can be used with any HTTP client.
+
+## 3.2.1
+
+### Patch Changes
+
+**[FIXED]** `Content-Length` header will no longer be appended to requests if its value is "0" - such situation caused Next.js server to crash
+
+## 3.2.0
+
+### Minor Changes
+
+- **[ADDED]** `logger` option to the `middlewareModule` config. This option allows you to turn on/off the logging of the SDK requests and responses or to provide a custom logger function.
+
+```diff
+import { initSDK, buildModule, middlewareModule } from "@vue-storefront/sdk";
+import { Endpoints } from "@vsf-enterprise/sapcc-api";
+
+const sdk = initSDK({
+  commerce: buildModule(middlewareModule<Endpoints>, {
+    apiUrl: "http://localhost:8181/commerce",
++   logger: true,
+  }),
+});
+```
+
+Logger can be also turned on by setting the `ALOKAI_SDK_DEBUG` environment variable to `true`.
+
+## 3.1.1
+
+### Patch Changes
+
+- **[CHANGED]** handle error for not available middleware
+
+## 3.1.0
+
+### Minor Changes
+
+- **[CHANGED]** cdnCacheBustingId is now optional
+
 ## 3.0.0
 
 ### Major Changes